% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.birthsbymaternalage}
\alias{data.birthsbymaternalage}
\title{Number of births by age of mother}
\format{A data frame with 20580 observations on the following 8 variables.
\describe{
  \item{country_code_numeric}{a numeric vector}
  \item{country_code}{country code}
  \item{country}{name of country}
  \item{age_from}{starting age}
  \item{age_to}{end age}
  \item{year}{year}
  \item{gender}{gender}
  \item{value}{number of births}
}}
\source{
{Demographic data from the Vaccine Impact Modelling Consortium (VIMC)}
}
\usage{
data.birthsbymaternalage
}
\description{
Dataset with the number of births by maternal by country, year and gender
}
\keyword{datasets}
