% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.birthsquinquennial}
\alias{data.birthsquinquennial}
\title{Number of births - quinquennial}
\format{A data frame with 2940 observations on the following 8 variables.
\describe{
  \item{country_code_numeric}{a numeric vector}
  \item{country_code}{country code}
  \item{country}{name of country}
  \item{age_from}{starting age}
  \item{age_to}{end age}
  \item{year}{year}
  \item{gender}{gender}
  \item{value}{number of births occuring every 5 years}
}}
\source{
{Demographic data from the Vaccine Impact Modelling Consortium (VIMC)}
}
\usage{
data.birthsquinquennial
}
\description{
Dataset containing number of births occuring every 5 years by country, age range, year and gender
}
\keyword{datasets}
