% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.disabilityWeights}
\alias{data.disabilityWeights}
\title{Disability weights for RV, HPV, Hib and Sp}
\format{An object of class \code{data.frame} with 19 rows and 6 columns.}
\source{
{Global, regional, and national incidence, prevalence, and years lived
with disability for 354 diseases and injuries for 195 countries, 1990–2017:
a systematic analysis for the Global Burden of Disease Study 2017}

\url{http://ghdx.healthdata.org/record/global-burden-disease-study-2017-gbd-2017-disability-weights}
}
\usage{
data.disabilityWeights
}
\description{
Data: Disability weights for Rotavirus (RV), Human papillomavirus (HPV),
Haemophilus influenzae type b (Hib) and Streptococcus pneumoniae (Sp).
The weights are measured on a scale from 0 to 1, where 0 equals a state of
full health and 1 equals death.
}
\keyword{datasets}
