% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.Hib_DALY}
\alias{data.Hib_DALY}
\title{Disability weights for Haemophilus influenzae type B}
\format{A data frame with 24 observations on the following 3 variables.
\describe{
  \item{Disease}{disease}
  \item{Condition}{condition/sequelae}
  \item{GBD_2015_mean}{Mean disability weight}
}}
\source{
{General Guidance for DALYs calculation VIMC with input from DOVE 2017-11-24 11:03:46}
}
\usage{
data.Hib_DALY
}
\description{
Dataset disability weights for Haemophilus influenzae type B by condition/sequelae
}
\keyword{datasets}
