% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.Hib_Hibmeningitis_deaths}
\alias{data.Hib_Hibmeningitis_deaths}
\title{Event rates (deaths) for Haemophilus influenzae type b (Hib): Hib meningitis}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{MCEE (Wahl et al 2017)}
}
\usage{
data.Hib_Hibmeningitis_deaths
}
\description{
Dataset containing event rates (deaths) for Hib meningitis
}
\keyword{datasets}
