% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.PCV_nonsevere_SpNPNM_visits}
\alias{data.PCV_nonsevere_SpNPNM_visits}
\title{Event rates (visits) for Streptococcus pneumoniae (Sp): Sp NPNM (non-severe)}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{DHS/MICS stratum}
}
\usage{
data.PCV_nonsevere_SpNPNM_visits
}
\description{
Dataset containing event rates (visits) for Sp NPNM (non-severe)
}
\keyword{datasets}
