% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.PCV_Spmeningitis_cases}
\alias{data.PCV_Spmeningitis_cases}
\title{Event rates (cases) for Streptococcus pneumoniae (Sp): Sp meningitis}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{MCEE (Wahl et al 2017)}
}
\usage{
data.PCV_Spmeningitis_cases
}
\description{
Dataset containing event rates (cases) for Sp meningitis
}
\keyword{datasets}
