% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.pdeathbyage}
\alias{data.pdeathbyage}
\title{Probability of dying by age}
\format{A data frame with 52920 observations on the following 8 variables.
\describe{
  \item{country_code_numeric}{a numeric vector}
  \item{country_code}{country code}
  \item{country}{name of country}
  \item{age_from}{starting age}
  \item{age_to}{end age}
  \item{year}{year}
  \item{gender}{gender}
  \item{value}{probability of dying}
}}
\source{
{Demographic data from the Vaccine Impact Modelling Consortium (VIMC)}
}
\usage{
data.pdeathbyage
}
\description{
Dataset containing probability of dying by age by country, age range, year and gender
}
\keyword{datasets}
