% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.quinquennialpop}
\alias{data.quinquennialpop}
\title{Quinquennial population (5-year time and age)}
\format{A data frame with 71442 observations on the following 8 variables.
\describe{
  \item{country_code_numeric}{a numeric vector}
  \item{country_code}{country code}
  \item{country}{name of country}
  \item{age_from}{starting age}
  \item{age_to}{end age}
  \item{year}{year}
  \item{gender}{gender}
  \item{value}{population quinquennial 5-year time and age}
}}
\source{
{Demographic data from the Vaccine Impact Modelling Consortium (VIMC)}
}
\usage{
data.quinquennialpop
}
\description{
Dataset containing population - quinquennial population (5-year time and age) by country, age range, year and gender
}
\keyword{datasets}
