% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.remainingyearsoflife}
\alias{data.remainingyearsoflife}
\title{Expected remaining years of life}
\format{A data frame with 64680 observations on the following 8 variables.
\describe{
  \item{country_code_numeric}{a numeric vector}
  \item{country_code}{country code}
  \item{country}{name of country}
  \item{age_from}{starting age}
  \item{age_to}{end age}
  \item{year}{year}
  \item{gender}{gender}
  \item{value}{expected remaining years of life}
}}
\source{
{Demographic data from the Vaccine Impact Modelling Consortium (VIMC)}
}
\usage{
data.remainingyearsoflife
}
\description{
Dataset containing expected remaining years of life by country, age range, year and gender
}
\keyword{datasets}
