% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.rotavirus_intussusception_deaths}
\alias{data.rotavirus_intussusception_deaths}
\title{Event rates (deaths) for Rotavirus intussusception}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{Clark et al, unpublished update of the Jiang literature review.
Medians of available country datasets within countries were used. For countries without data,
the median for all datasets in the same WHO region was used. This was a better indicator of incidence
 than U5MR quintile e.g. WPRO has uniquely high incidence.}
}
\usage{
data.rotavirus_intussusception_deaths
}
\description{
Dataset containing event rates (deaths) for rotavirus intussusception
}
\keyword{datasets}
