% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.rotavirus_nonsevere_RVGE_visits}
\alias{data.rotavirus_nonsevere_RVGE_visits}
\title{Event rates (visits) for Rotavirus non-severe RVGE}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{Assume 1 visit for every 2 non-severe cases}
}
\usage{
data.rotavirus_nonsevere_RVGE_visits
}
\description{
Dataset containing event rates (visits) for rotavirus D1 for non-severe RVGE
}
\keyword{datasets}
