% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.rotavirus_severe_RVGE_cases}
\alias{data.rotavirus_severe_RVGE_cases}
\title{Event rates (cases) for Rotavirus severe RVGE}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{Fischer-Walker C. et al, Table 1: Global and regional burden of diarrhoea and pneumonia per
year in children aged 0–4 years, by WHO region.  Global burden of childhood pneumonia and diarrhoea.
Lancet 2013; 381: 1405–16.  Notes: Episodes per child per year <5yrs (2010) by WHO region were multiplied by
the proportion of episodes that were severe by WHO region (approximately 2%).  These were further multiplied
by the rotavirus-positive proportion <5yrs by WHO region, reported among hospitalised diarrhoea cases in CHERG
(Lanata C. et al, Global Causes of Diarrheal Disease Mortality in Children <5 Years of Age: A Systematic Review.
PLOS One. September 2013, Volume 8, Issue 9, e72788).  The uncertainty range only reflects uncertainty in the
incidence of diarrhoea episodes.}
}
\usage{
data.rotavirus_severe_RVGE_cases
}
\description{
Dataset containing event rates (cases) for rotavirus D2 for severe RVGE
}
\keyword{datasets}
