% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.rotavirus_severe_RVGE_hosps}
\alias{data.rotavirus_severe_RVGE_hosps}
\title{Event rates (hospital) for Rotavirus severe RVGE}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{See Clark et al, TRIVAC, Vaccine, Appendix}
}
\usage{
data.rotavirus_severe_RVGE_hosps
}
\description{
Dataset containing event rates (hospitals) for rotavirus for severe RVGE
}
\keyword{datasets}
