% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.rotavirus_severe_RVGE_visits}
\alias{data.rotavirus_severe_RVGE_visits}
\title{Event rates (visits) for Rotavirus severe RVGE}
\format{A data frame with 195 observations on the following 8 variables.
\describe{
  \item{Country}{country}
  \item{WHO region}{WHO region}
  \item{WHO region2}{WHO region country code}
  \item{Income}{income level}
  \item{Age}{age}
  \item{Mid}{mid}
  \item{Low}{low}
  \item{High}{high}
}}
\source{
{Clark A et al, unpublished update of the IHME/MCEE/WHOCDC estimates. Median/Min/Max 2015 estimates
were used or the estimate for the most recent pre-vax year (using WUENIC 15th July 2017). }
}
\usage{
data.rotavirus_severe_RVGE_visits
}
\description{
Dataset containing event rates (visits) for rotavirus for severe RVGE
}
\keyword{datasets}
