% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.survival}
\alias{data.survival}
\title{Survivors from a birth-cohort of 100k}
\format{A data frame with 20580 observations on the following 8 variables.
\describe{
  \item{country_code_numeric}{a numeric vector}
  \item{country_code}{country code}
  \item{country}{name of country}
  \item{age_from}{starting age}
  \item{age_to}{end age}
  \item{year}{year}
  \item{gender}{gender}
  \item{value}{number of survivors}
}}
\source{
{Demographic data from the Vaccine Impact Modelling Consortium (VIMC)}
}
\usage{
data.survival
}
\description{
A data frame with 64680 observations on the following 8 variables.
}
\keyword{datasets}
