% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.totalpop}
\alias{data.totalpop}
\title{Total population}
\format{A data frame with 14798 observations on the following 8 variables.
\describe{
  \item{country_code_numeric}{a numeric vector}
  \item{country_code}{country code}
  \item{country}{name of country}
  \item{age_from}{starting age}
  \item{age_to}{end age}
  \item{year}{year}
  \item{gender}{gender}
  \item{value}{total population}
}
# @source {Demographic data from the Vaccine Impact Modelling Consortium (VIMC)}}
\usage{
data.totalpop
}
\description{
Dataset containing total population country, age range, year and gender
}
\keyword{datasets}
