% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.vaccine_schedules}
\alias{data.vaccine_schedules}
\title{Disability weights for Haemophilus influenzae type B}
\format{A data frame with 195 observations on the following 6 variables.
\describe{
  \item{Country}{country}
  \item{BCG}{BCG target age in weeks}
  \item{DTP1}{DTP1 target age in weeks}
  \item{DTP2}{DTP2 target age in weeks}
  \item{DTP3}{DTP3 target age in weeks}
  \item{Meas1}{Measles target age in weeks}
}}
\source{
\url{www.who.int/immunization/monitoring_surveillance/data/schedule_data.xls}
}
\usage{
data.vaccine_schedules
}
\description{
Dataset containing disability weights for Haemophilus influenzae type B by condition/sequelae (UNIVAC model)
}
\keyword{datasets}
