% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_vimr.R
\name{EstimateVaccineImpact}
\alias{EstimateVaccineImpact}
\title{Estimate vaccination impact and generate disease burden estimates}
\usage{
EstimateVaccineImpact(vaccineCoverageFile, diseaseBurdenTemplateFile,
  plots = FALSE)
}
\arguments{
\item{plots}{}
}
\description{
Estimate vaccination impact and generate disease burden estimates
}
