% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotVaccineImpact_vimc}
\alias{PlotVaccineImpact_vimc}
\title{Plot vaccine coverage and disease burden estimates}
\usage{
PlotVaccineImpact_vimc(vaccineCoverageFile = NULL,
  diseaseBurdenTemplateFile, diseaseBurdenEstimatesFile = NULL)
}
\arguments{
\item{diseaseBurdenEstimatesFile}{}
}
\description{
Plot vaccine coverage and disease burden estimates
}
