% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_vimr.R
\name{SaveDieaseBurdenEstimates}
\alias{SaveDieaseBurdenEstimates}
\title{Save disease burden estimates in full and streamlined VIMC format}
\usage{
SaveDieaseBurdenEstimates(diseaseBurdenEstimates,
  diseaseBurdenTemplateFile)
}
\arguments{
\item{diseaseBurdenTemplateFile}{}
}
\description{
Save disease burden estimates in full and streamlined VIMC format
}
