% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_vimr.R
\name{SetupCohorts}
\alias{SetupCohorts}
\title{Generate cohorts file based on vaccine coverage and disease burden template
files (VIMC format)}
\usage{
SetupCohorts(vaccine, vaccineCoverageFile, diseaseBurdenFile)
}
\arguments{
\item{diseaseBurdenFile}{}
}
\description{
Generate cohorts file based on vaccine coverage and disease burden template
files (VIMC format)
}
