% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{univacr}
\alias{univacr}
\alias{univacr-package}
\title{univacr: UNIVAC decision support model}
\description{
A universal framework for evaluating vaccine policy options in low- and
middle-income countries. This R package is based of the spreadsheet-based
tool that is accessible at https://www.paho.org/provac-toolkit/.
}
\section{Vaccine impact model}{

\describe{
  \item{Estimate the health impact and cost-effectiveness of vaccination.}
  \item{Hib, pneumococcal and rotavirus vacccination can be separately analysed.}
}
}

