% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vimr}
\alias{vimr}
\alias{vimr-package}
\title{vimr: Vaccine Impact Model}
\description{
"vimr" is a vaccine impact model (R package) for estimating the health impact
of vaccination programmes at the national level for a given set of countries.
This R package is based of the spreadsheet-based tool (UNIVAC) which is a
universal framework for evaluating vaccine policy options in low- and
middle-income countries, and is accessible at
https://www.paho.org/provac-toolkit/,
}
\section{Vaccine impact model}{

\describe{
  \item{Estimate the health impact of vaccination at the national level.}
  \item{Hib, pneumococcal and rotavirus vacccination can be separately analysed.}
  \item{Multiple countries can be analysed and vaccine impact estimates are generated for each country.}
}
}

