% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AdjustDiseaseBurden}
\alias{AdjustDiseaseBurden}
\title{Update disease burden estimates to the timeline specific to this cohort}
\usage{
AdjustDiseaseBurden(diseaseBurden)
}
\arguments{
\item{diseaseBurden}{}
}
\description{
Update disease burden estimates to the timeline specific to this cohort
}
