% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{univacr}
\alias{univacr}
\alias{univacr-package}
\title{univacr: UNIVAC decision support model}
\description{
UNIVAC is a UNIversal framework for evaluating VACcine policy options in low- and middle-income countries. The R package is based of the spreadsheet-based tool (see http://www.paho.org/provac-toolkit/).
}
\section{UNIVAC is based of PROVAC}{

\describe{
  \item{PAHO’s ProVac Initiative helps national immunization programs managers and policymakers make more robust, evidence based decisions.}
  \item{Since 2004, the ProVac Initiative and its collaborators have trained hundreds of people at ministries of health in the Americas and elsewhere on the development and use of vaccine impact, cost and cost-effectiveness data to inform decision making.}
  \item{More than 30 countries have received direct technical assistance to conduct cost or cost-effectiveness analyses on immunization programs from the Initiative.}
}
}

